First piece of serviceable code
So It turns out that when I created the project, I created it as a unit
test instead of a real project.  I fiddled with it for a while but in
the end, opted to let VS create a fresh solution for me and then copied
in Character.cs, Monster.cs and Project.cs.

-Character.cs is an abstract class that handles all characters, their
stats and basic interactions, currently including: taking damage and
displaying the HP value.
-Monster.cs currently contains base initialised stats and allows me to
edit the name while instantiating it.
-Program.cs features a simple test of showing the HP, dealing damage and
then showing the HP again.
-UIUpdate.cs is my way of segregating all of the UI code from the game
logic.  Currently it features an overloaded function to write data to
the console.