End of day 1
So I've implemented a menu system using methods in UIUpdate however I'm
not sure if the menu code shouldn't be in a class of its own.  This menu
system relies on a switch statement in Program to decide what to do with
it.

I have highlighted a few things to work on to make the code both more
readable and flexible.