New Files, new approach
I have deleted all of my previous files in favour of three simple
classes: Program, Character and Technique.

Program is now empty as I am now performing all of the initial tests
within unit tests.
Technique is essentially empty, save two method shells.

Monster is where the bulk of the work happens.  So far I have huge huge
sections of commented out code.  This code is one approach that I
decided would be cumbersome in its implementation.  I have kept the code
so that I can get a second opinion as to which is the best approach.  On
top of this, all of my classes are public.  This will continue until I
have a working prototype of an automated battle between two Characters,
after which point, I will tweak the access modifiers and pull back to
the main branch.

For now, the way the code runs is this:
1) Create a dictionary of baseStats
2) Create a dictionary of Characters
3) Initialise the Characters using baseStats.
4) Use Methods within Character to calculate the complete Stat profile
using the base stats as material.

Next I will implement the Technique Class.