Affinity Implemented
I have created an affinity feature so that when you instantiate a
character, you give it it's training points for the most basic stats
(peripheral vision, main vision, hearing, etc.) and it multiplies them
by the affinity for that stat tree (in this case, perception).

This means that certain characters have yield greater results from
training these categories.