Lots of changes while figuring out attacks
This update took a lot of experimenting as I was trying to work out how
to dynamically create attacks.  I spent quite a while looking at the
Strategy pattern but it would require me to hard code each attack in the
game, which I'm not sure I want to do.

I used an online tutorial to help me utilise the decorator pattern to
create a simple attack decorator within DecoratorTest.cs.  This is only
a working example and my next goal is to create a decorator system that
integrates with the game's stat systems.

I fixed a few capitalisation mistakes throughout the code and replaced
every integer variable and every integer array with doubles and double
arrays as this is more useful when performing calculations.

I removed lightning from the core elements in the affinitylist and
changed wind to air.

I've created a list of all the elements, includng sub elements (i.e.
water + air = ice).

I then added element mastery to the character classes/sub-classes